Python/requirements.txt
math
scipy.optimize
pandas
matplotlib.pyplot